delete

+++ app/Http/Controllers/AccountingSystem/AccountController.php
@@ -170,10 +170,19 @@ class AccountController extends Controller
      */
     public function destroy($id)
     {
-        $account =AccountingAccount::findOrFail($id);
-        $account->delete();
-        alert()->success('تم حذف  الحساب بنجاح !')->autoclose(5000);
-        return back();
+        $entries=AccountingEntryAccount::where('account_id',$id)->get();
+
+        if($entries->count()==0){
+            $account =AccountingAccount::findOrFail($id);
+            $account->delete();
+            alert()->success('تم حذف  الحساب بنجاح !')->autoclose(5000);
+        }else{
+            alert()->warning('لا يمكن حذف الحساب  لوجود قيود   !')->autoclose(5000);
+
+        }
+
+        return redirect()->route('accounting.ChartsAccounts.index');
+
     }
 
     public  function  active($id){
